refactor: 알림을 배틀 알림으로, 알림 반응을 배틀 알림 반응으로 변경
