+++ src/main/java/com/poolex/poolex/alarm/battlealarm/domain/BattleAlarmType.java
@@ -1,0 +1,9 @@
+package com.poolex.poolex.alarm.battlealarm.domain;
+
+public enum BattleAlarmType {
+    EXPENDITURE_CREATED,
+    EXPENDITURE_NEEDED,
+    OVER_BUDGET,
+    ZERO_EXPENDITURE,
+    BATTLE_NOTIFICATION_CHANGED
+}
