--- src/main/java/com/poolex/poolex/alarm/domain/AlarmType.java
@@ -1,9 +1,0 @@
-package com.poolex.poolex.alarm.domain;
-
-public enum AlarmType {
-    EXPENDITURE_CREATED,
-    EXPENDITURE_NEEDED,
-    OVER_BUDGET,
-    ZERO_EXPENDITURE,
-    BATTLE_NOTIFICATION_CHANGED
-}
