--- src/main/java/com/poolex/poolex/alarmreaction/domain/AlarmReactionContent.java
@@ -1,39 +1,0 @@
-package com.poolex.poolex.alarmreaction.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-import org.springframework.util.StringUtils;
-
-@Embeddable
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AlarmReactionContent {
-
-    private static final int MIN_LENGTH = 2;
-    private static final int MAX_LENGTH = 30;
-    @Column(name = "content")
-    private String value;
-
-    public AlarmReactionContent(final String value) {
-        validate(value.strip());
-        this.value = value.strip();
-    }
-
-    private void validate(final String value) {
-        if (!StringUtils.hasText(value)) {
-            throw new IllegalArgumentException();
-        }
-        validateLength(value);
-    }
-
-    private void validateLength(final String value) {
-        if (value.length() < MIN_LENGTH || value.length() > MAX_LENGTH) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public String getValue() {
-        return value;
-    }
-}
