--- src/main/java/com/poolex/poolex/alarmreaction/domain/AlarmReactionRepository.java
@@ -1,7 +1,0 @@
-package com.poolex.poolex.alarmreaction.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AlarmReactionRepository extends JpaRepository<AlarmReaction, Long> {
-
-}
