--- src/main/java/com/poolex/poolex/alarmreaction/domain/AlarmReactionType.java
@@ -1,15 +1,0 @@
-package com.poolex.poolex.alarmreaction.domain;
-
-import java.util.Arrays;
-
-public enum AlarmReactionType {
-    PRAISE,
-    SCOLD;
-
-    public static AlarmReactionType findByName(final String name) {
-        return Arrays.stream(values())
-            .filter(reactionType -> reactionType.name().equals(name.toUpperCase()))
-            .findFirst()
-            .orElseThrow(IllegalArgumentException::new);
-    }
-}
