--- src/main/java/com/poolex/poolex/alarmreaction/service/dto/request/AlarmReactionCreateRequest.java
@@ -1,16 +1,0 @@
-package com.poolex.poolex.alarmreaction.service.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class AlarmReactionCreateRequest {
-
-    private Long alarmId;
-    private String type;
-    private String content;
-}
