+++ src/main/java/com/poolex/poolex/battlealarmreaction/domain/BattleAlarmReactionRepository.java
@@ -1,0 +1,7 @@
+package com.poolex.poolex.battlealarmreaction.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BattleAlarmReactionRepository extends JpaRepository<BattleAlarmReaction, Long> {
+
+}
