+++ src/main/resources/sql/schema.sql
@@ -80,19 +80,22 @@ create table battle_notification
     primary key (id)
 );
 
-create table alarm
+create table battle_alarm
 (
     id         bigint generated by default as identity,
     battle_id  bigint,
     member_id  bigint,
     type       varchar(255) check (type in
-                                   ('EXPENDITURE_CREATED', 'EXPENDITURE_NEEDED', 'OVER_BUDGET',
-                                    'ZERO_EXPENDITURE', 'BATTLE_NOTIFICATION_CHANGED')),
+                                   ('EXPENDITURE_CREATED',
+                                    'EXPENDITURE_NEEDED',
+                                    'OVER_BUDGET',
+                                    'ZERO_EXPENDITURE',
+                                    'BATTLE_NOTIFICATION_CHANGED')),
     created_at timestamp(6),
     primary key (id)
 );
 
-create table alarm_reaction
+create table battle_alarm_reaction
 (
     id         bigint generated by default as identity,
     alarm_id   bigint,
